requirements.txt
chromadb>=0.5.0
sentence-transformers>=2.2.2
pyyaml>=6.0
tqdm>=4.66.0
rich>=13.7.0
ollama>=0.3.0
PyPDF2>=3.0.0
python-docx>=0.8.11
